Revert "pmt-ptw"
This reverts commit 6a28471d857ed9fc2ff006b3021792dde658fb3b.

+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/bag/sorted/immutable/ImmutableSortedBagImpl.java
@@ -103,7 +103,7 @@ class ImmutableSortedBagImpl<T>
             size += occurrence;
         }
 
-        this.size = ((Integer)(size)).intValue();
+        this.size = size;
     }
 
     @Override
